Made the backend more organized again.
