+++ backend/.eslintrc.js
@@ -12,6 +12,7 @@ module.exports = {
   rules: {
     "indent": ["error", 4],
     "@typescript-eslint/indent": ["error", 4],
+    "@typescript-eslint/explicit-function-return-type": "off",
     "semi": ["error", "always"],
     "@typescript-eslint/semi": ["error", "always"],
     "comma-dangle": ["error", {
